requirements.txt
asyncio==3.4.3
selenium==4.21.0
selenium-wire==5.1.0
webdriver-manager==4.0.1
websocket-client==1.8.0